окно открытия при успешной загрузке
