+++ js/main.js
@@ -3,7 +3,6 @@ import './server.js';
 import './utils/get-random-positive-integer.js';
 import './create-card.js';
 import'./create-small-picrure.js';
-import'./create-big-picrure.js';
 import './form.js';
 import './form-upload.js';
 import './open-picrure.js';
