+++ js/server.js
@@ -1,23 +1,20 @@
-const createLoader = (onSuccess, onError) => () => fetch(
-  'https://24.javascript.pages.academy/code-and-magick/data', {
-    method: 'GET',
-    credentials: 'same-origin',
-  },
-)
-  .then((response) => {
-    if (response.ok) {
-      return response.json();
-    }
 
-    throw new Error(`${response.status} ${response.statusText}`);
-  })
-  .then((data) => {
-    onSuccess(data);
-  })
-  .catch((err) => {
-    onError(err);
-  });
+const createLoader = (onSuccess, onError) => () => {
+  return fetch(
+    'https://24.javascript.pages.academy/kekstagram/data',
+    {
+      method: 'GET',
+      credentials: 'same-origin',
+    },
+  )
+    .then((response) => {
+      if (response.ok) {
+        return response.json();
+      }
+        throw new Error(`${response.status} ${response.statusText}`);
+    })
+};
 
 export {
   createLoader
-};
+}
